WIP
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@2422 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ build/checkversion.sh
@@ -1,5 +1,9 @@
 #!/bin/sh
 
+if [ -f .noversion ] ; then
+    exit
+fi
+
 if eval test x${1} = xforce ; then
 	rm -f .version
 fi
